requirements.txt
flask
pillow
hashlib